[4.0] admin menu reduced motion (#33069)
pr for #31242

This is a temp fix until upstream addresses this https://github.com/onokumus/metismenujs/issues/9

The fix applied is the exact same one used by webaim.org so should be acceptable from an a11y perspective.

There will be no change for most users. It is only if they have reduced motion set in their operating system that they encounter the problem this fixes.