[4.0] webp support in templates
#31048 provided the library level support for webp images

This PR adds webp to the list of allowed image types for com_templates. It does not have an update sql as that would change user data.

This isnt really anything to test here accept codestyle as using webp images in templates was the test for the libray update so we know it works