[4.0] Default Menu item (#33124)
When setting the default menu item on and off you will see that the publish up and publish down fields are shown/hidden as appropriate. However because they are above the home button it creates a usability and an accessibility as you have already navigated passed those fields.

This PR simply changes the order.