[4.0] Fix tinymce template
TinyMCE has a really powerful feature that almost knows about - the ability to have predefined snippets of text and html.

This is not a new feature of TinyMCE it has been present for at least 10 years. The reason no one knows about this or talks about it is that its very hard to use.

I am only aware of a blog post by @coolcatcreations explaining how to use it and you will see it is both very powerful but also a pain to setup

## The Problem
The folder that tinymce looks in for the stored "templates" is outside of the folder structure that can be accessed within the Joomla adminsitrator ui. In Joomla 4 it is `media/vendor/tinymce/templates` so it can only be accessed by ftp or cli.

## The Solution
Change the folder where the "templates" are stored to a user defined folder in the site/templates folder structure. This enables the user to use the template manager to create and edit the "templates".

## Backwards Compatibility
No problem here as if a custom folder is not set then the default is the current `media/vendor/tinymce/templates`. The one difference is that the current (and therefore default) behaviour is for the files to be *.html. Because that filetype is by default prohibited in the template manager I have changed it to the supported .txt filetype __only__ when using the custom folder/

### Bonus
A template developer can include predefined snippets of text and html with their template

## To Test
Firstly familiarise yourself with this TinyMCE feature. You will find it on the Insert dropdown of the editor toolbar. TinyMCE ships with two examples you can play with.

### Apply the PR
Go to the Editor - TinyMCE plugin and there is a new field where you can select the folder to use.
Make sure you are in a configuration set that applies to your userlevel (it is probably Set 0) and then select a folder and save.
Go to an article and try to insert a template (as you did before)
You should get a template not found message
Go to the template manager and create two text files in the folder you previously selected
- with just plain text
- with basic html
Now return to the article and you should be able to insert both these templates successfully.

## Issues to solve - need help
The list of folders presented to the user to select from was intended to be limited to templates\[template_name]\html\ and all its subfolders but despite my best efforts I was not able to get the folderFilter to includes those subfolders