[4.0] Sample Data menu permission (#33299)
The permission for the template settings menu item was set to public. This meant that as long as you could log in to the site in the front end then you would see the link to edit the template. However only a super user can do this so you got an ugly error page.

This PR correctly sets it to Super Users the same as with the Site settings menu itm