[4.0] Empty Trash Button
When you have filtered your items by the trashed status in most components there is an empty trash button on the toolbar. However for contacts, newsfeeds, redirects and tags the button is in the actions dropdown.

For a consistent UI they should all be the same.

To test apply the pr
filter a component by trashed and observe the button