[4.0] TinyMCE plugin layout (#33177)
* [4.0] (WIP) TinyMCE plugin layout

There have been multiple issues raised about the UI of the tinymce plugin and the lack of clarity that the options are for different sets. It's also a very regular topic on the forum whereby users believe they have configured the plugin and it doesn't work. When the problem is that they are simply in the wrong set.

This is a work in progress attempt to improve this.

The basic concept of this change it to move the "All available menus and buttons." inside each set so that the tabs are at the top.

The other main change (which is not done yet) is to move the "Assign this Set to" field and label to the very first field in the tab.

Hopefully this will make it much much more obvious that you are completing a set of options for a usergroup and not for the entire plugin.

(If anyone has a suggestion how to display that field at the very top it would be appreciated)

* access field

* draft

* date