[4.0] Debug Group scope
The attribute scope is essential to tell users of assistive technology if the header applies to the column or the row. However that means it can only be applied to a table header (th) and not to a table cell (td).

There is a small cosmetic change as a result in that the contents of the cell are now visuall displayed in bold.

To test go to users-> groups and click on the permissions icon.

All the content in the first column should be in a th with a scope of row and visually displayed as bold