[J4-Template] Add position without parent tag
**`<jdoc:include type="modules" name="top" style="default" class="grid-child" />`**

and

**`<jdoc:include type="modules" name="bottom" style="default" class="grid-child" />`**

Friends, please consider the offer to add 2 more positions to the template.
What is it for?
Answer: These fields will have little value for users who use the Cassiopeia template as it is.
BUT!, While a lot of developers use the Cassiopeia template as a basis for redefining new styles. I do it myself.
The fact is that the Bootstrap is already built into Cassiopeia, the grid, controlls, margins, and overrides of default styles are already configured. You don't need to do anything about it. Therefore, it is logical for any templates to use a ready-made Cassiopeia template to redefine its styles.
This is to add a background image, change the color. Simple as that.
But at the same time, you need to have your entire position without the parent tag.
so that you can place as many modules as you want in this position, each of which will have its own grid for mobile screens.
```
<?php if ($this->countModules('bottom-b')): ? >
	<div class="grid-child container-bottom-b">
		<jdoc:include type="modules" name="bottom-b" style="card" />
	</div>
<?php endif; ?>
```
The container-bottom-b class and any other container class have their own built-in grid. What causes problems when the grid is not needed at all in the container class.

`<jdoc:include type="modules" name="bottom" style="default" class="grid-child" />`

**`full-width`** or **`containier-full`** (for old template as 'protostar') of **`wrapper-fluid`**

In addition, I suggest that you be able to use your own class in `<jdoc:include type="modules" name="bottom" style= "default" css="grid-child"/>`.
Thus, the position will be automatically placed in the template grid in the center of the content.
However, if the class prefix is added by the user, I suggest excluding the position class altogether. This will allow the user to add the full-grid class themselves, or to add the `grid` class or `card` class themselves.

for example, you need to make a map at the bottom of the site in the full width of the site with the route. But how to do this if the container is bounded by the central column of the grid.