[4.0] revert removal of background
In the repaint the background image was removed and this has had the side effect of breaking the ability to make the screen monochrome. This PR reverts that change.

To test apply the pr and rebuild the css

Then go to the template admin styles for admin and set the color to monochrome

Before this PR the background stays blue
With this PR the background is monochrome