[4.0] com_media drag and drop
The dropzone is the height of the media-browser div. Before this PR that height was dependent on the height of the content. Now it is the full height of the div.

its a css change so npm i

## Test
The easiest way is to create a new folder in the media manager and try to upload a new file using dragdrop.

### Before

### After