[4.0] Header height atum (#33756)
Buttons have a padding of 1px defined in the browser unless it is specified in the css

On the toolbar they might all look like buttons but only the lastone (user) is a button. Therefore that one is 68px and the others are 66px.

This results in a 2px difference in the height of the header as shown in the reported issue.

This PR removes the browser padding top and bottom and therefore the header is always 66px high.