[4.0] Revert #33839 (#33865)
It is absolutely not necessary to change titles to tooltips. It is a pointless exercise with zero benefits that would need to be done in a gazillion places.

The only time we should be using the black tooltips is as a replacement for the bootstrap tooltips. They are NOT a replacement for titles.