[4.0] metatitle
@philetaylor spotted this #

Meta tags are for data that cannot be represented by html. [mdn](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meta) specifically states that title is one of the things you do NOT need to set a meta tag for.

I am guessing that at some point over 10 years ago we mistakenly had this tag and it was removed but this one bit was left.

This PR stops the value being set in configuration.php on new installs
This PR stops a frontend display of a newsfeed adding meta name=title to the head
This PR will not remove the line from an existing configuration.php as its userdata we shouldnt touch. But as its no longer referenced anywhere there is no problem.