[4.0] joomla pre-update component improvements
This is just a start of the improvements that are required to make the joomla pre-update check page formatted correctly and accessible.

In order to test this page you will need to force joomla to see an update. One way to do this is to set the update channel to custom and use this url https://update.joomla.org/core/nightlies/next_major_list.xml

When you have done that then you will see the pre-update page that this PR is addressing.

The changes addressed in this PR
1. removes inline style used to set the size of the > icon and uses css class instead
2. adds space between the title and the [

There is much more to do but this will be addressed in further PR (hopefully by more people than me) so PLEASE don't comment on any other issue on this page - they will be ignored.