[4.0] List contacts in a category
This view was not checking correctly to see if it should output anything for a field. As a result there were a lot of empty rows

To test create a few contacts with various amounts of detail
Create a menu item of type List contacts in a category

When viewed on the frontend you will see numerous blank lines in the details

Apply the pr and those blank lines are remmoved.