[4.0] Admin modules
As Phil discovered #33520 several of the admin modules really dont work well in a position other than the position they were designed for. It doesn't make sense to write code to resolve this if it is even possible. The mod_version description said ` and is intended to be displayed in the 'status' position.` So I just added that to the description of the other modules that this applies to.