[4.0] Subcategories heading
Content and Contacts have an option labeled "Subcategories Text"

What it does is to add the h3 heading "Subcategories" above the list of subcategories

This PR changes the text to be "Subcategories Heading" which I believe is a more accurate description.

Because the key for the language string is a global string and misnamed as JGLOBAL_SHOW_CATEGORY_HEADING_TITLE_TEXT_LABEL and not SUBcategory I added a new string.

The name of the field, also misnamed as show_category_heading_title_text, has not been changed as that would break b/c for no good reason.

Finally, I suspect because of the incorrect name, the option was not being displayed in the correct, and consistent, place in the form and this PR fixes that as well.

There is no change to anything on the frontend