[4.0] Notification icons
The order of the notification icons on the dashboard is determined by the order of the quickicon plugins

This PR changes the order based on importance - so that Joomla Update is first.

Note that phpversioncheck is last because it is not displayed in the notifications section but in the messages.

Note this is for new installs only as a user could have customised the ordering themselves