[4.0] pre-update check [a11y] [ui]
In the required section of the pre-update check there are three requirements that have some extra information available if the requirment is not met.

Currently they are completely inaccessible as they are a title on a non-focusable element. There really isnt a good reason for this information not to be visible by default when it should be displayed