[4.0] com contact unused fields
The form for an individual contact defines fields for all the icon images so that you can override them at the contact level as well as the component level.

However the fields are never displayed so they are never used. I have gone back over 8 years in the history and I cant see them being used at any point in that time period.

This pr removes the fields from the xml. No point in maintaining code that is never used