[4.0] sysinfo th (#34406)
When this css was added the table headers of the template were not bold. Since they are now we don't need this extra css file.

Note that th in the header is still font-weight: 700 but the th in the body will be 500 - this is consistent with all other tables now.