[4.0] Dont try to translate missing label (#34412)
* [4.0] Dont try to translate missing label

There are several fields that are hidden and they do not have a label. Before this PR Joomla would use the field name as a label in this instance and then try to run it through JTEXT to translate it.

However the label is never rendered for these hidden fields so it is a pointless exercise AND it results in the language debugger reporting a missing language string.