[4.0] com_content set default date field
This PR lets you set the default date field to be used in lists. eg published, created, modified

Before this PR you could only set the field if you had also set a default list order that was date based, which didnt actually make sense.

It was also confusing as the use of showon for this field was only in the global config for com_content and not at the menu level.