[4.0] Content versions button [a11y]
Content versions for an article has a button to indicate if the version should be kept forever or not.

However it is incorrectly an `a` link designed to look like a button but not act like a button.

If the action is to take you to a new page then its a link - otherwise its a button.

A button can be activated with both the space and enter keys. A link just by the enter key.

BONUS: This also fixes #34592