[4.0] media manager toolbar
The toolbar in the media manager is sticky but it has no background. This simple PR adds a solid white background (just as it has in the cassiopeia template.

To test rebuild the css and make sure you have enough images in media manager to make it scroll.

You will see there is no background on the toolbar when you scroll and you can see the images underneath it

Apply this PR and repeat and you now have a solid background