[4.0] oops png
png was accidentally placed in the doc and the image filetypes. this pr corrects that.

to test you will need to apply the pr
go to the media manager options
delete the content of the Legal Document Extensions (File Types)
save
(this will put the correct defaults back)
go to articles and add a png as an intro image