[4.0] Linked Image (atum) a11y
When an image is used for functional purposes eg a link then then alt text should describe the function and not the image. Otherwise the link may appear to be empty to assistive technology and therefore ignored.

https://www.w3.org/WAI/tutorials/images/functional/#image-used-alone-as-a-linked-logo

In the atum template the logo in the top right is sometimes used as a link to the homepage. So in those circumstance it needs an alt text that describe where the link takes you that and not an aria label on the image.