[4.0] mod_languages css
This PR changes the css so that we don't need to have any RTL specific css as it now uses logical css properties.

To test you will need a multilingual site and have set the language switcher module to be displayed as a dropdown. After the pr there will be no visible change