[4.0] input-group-append prepend
This css class was removed in bootstrap 5 and we are not using the markup anywhere in the core so it makes no sense to keep it and keep maintaining css that isnt used.

To test I guess code review, make sure the scss still compiles and confirm that the classes are not being used.

NOTE: it is still used in plg_installer_webinstaller and that will be addressed in a sperate pr as it also requires upstream changes