In the message displayed after Switching the Administrator Language, show the Native Title instead of Title [in English]
For example, if German is Switched as the back-end Administrator language, following message is displayed:
Die Administrationssprache wurde auf "German (Germany)" geändert. 

With the change proposed, the following message would be displayed:
Die Administrationssprache wurde auf "Deutsch (Deutschland)" geändert.

When the message is displayed in the native language (e.g., German), it would be nice/better to show the language's Native Name/Title instead of the Name/Title in English.