[4.0] cog icon
Currently we are using the same cog-icon as both an illustrative image next to options and as a functional image for the mmodule dropdown.

To avoid confusion this PR changes the functional image to a slightly different icon.

NOTE: personally I would prefer if these buttons looked like buttons but thats beyond the scope of this PR