[4.0]Translate menus
On a site with multiple admin languages you would expect that the name of the menu could be translated.

## Test instructions
Create an admin menu and give it a title such as JADMINISTRATOR
Create a site menu and give it a title such as JSITE

## Expected behaviour
The names of both menus are translated

## Actual behaviour
Only the name of the admin menu is translated

## Solution
I have no idea why but there is an extra space at the end of `$item->title` so the string being passed through JText is "JSITE " which obviously doesn't match
I have no idea why there is not an extra space for the admin mmenu

This PR trims `$item->title` so that now the site menu name is also translated.

No idea if this is a hack/bandage or the correct solution but its been bugging me for hours and I can't find the space