[4.0] List Views (consistency)
In some list views the title cell contains additional information eg alias or category.
Sometimes they are on the same line, sometimes they are on new lines.

This PR makes them all on new lines for consistency. It is also much easier this way if you need to scan through a list looking for something.