The order of panel widgets in the system section
![image](https://user-images.githubusercontent.com/6898474/127853303-ed6658ab-beb2-4d56-8bab-84feb7ac9c46.png)

Colleagues, by what principle are the link panels in the system section sorted?
I would prefer the SETUP and MANAGE panels to be the first.
I feel tension every time I open a panel and try to find the right panel somewhere in the middle of other panels. And if the screen is small, then you will have to search for the most popular panel by rewinding the page.
.
Let's place the SETUP and MANAGE panels first.

I marked with numbers how I would like to see the order. But, I may be inaccurate. The main thing is that I would like to see the SETUP and MANAGE the in  first position.