Fix form manipulation by 'job' plugins
Before this change, job plugins were only loaded while displaying the
form. This meant it was not possible to save an injected form field
to the `params` column or change its value, since the injected XML did
not show itself while validating form data. This commit fixes this
issue, adds job plugins to some events in the constructor (needs review)
and adds the new trashed state to the STATES array.