[4.0] Update Information
By default bootstrap displays the caption for all tables below the actual table. Normally thats not an issue as they are also hidden and only used for screenreaders. However in the update component they are displayed and have important/useful information. You only really see why it a bad idea the information being at the end of the table is a problem when you are on a smaller screen or have a lot of data in the table.