Add CronjobPluginTrait for job plugins
Trait provides the common boilerplate associated with a job routine.
All a plugin needs to support the job trigger API is to use this trait
and call ::jobStart() and ::jobEnd() with the required parameters.
The ::jobEnd() method logs by default in the `cronjobs` category but
that behaviour can be disabled with a flag.