Bugfix form enhancement in CronjobPluginTrait
Adds primarily a fallback to Input in case neither the form nor the
data have a jobId to refer to.
This is needed because the way FormBehaviourTrait::loadForm() is, it
is not possible to:
(a) pass data to loadForm at all
(b) direct loadForm to load data, but not bind it.

This means that when plugins try to access data on submission, passed
to the model's getForm() method, they cannot as the onContentPrepareForm
event is fired by a non-context aware part of the process.
The workaround we opted for is, getting data through the request
manually. This is not perfectly ideal but just works.