Add a legal note
I reckon this is important to protect Open Source
Matters, Inc. My original work is licensed under the
GPLv3, this derivative I chose to license under the GPLv2
and share the copyright with OSM. If there is no
documentation about this having taken place, accepting
this PR could be a copyright violation. I'm not sure
since I am the author of both the original code and the
derivative but better be safe than sorry, right?