Fix task exit code, failure tracking
The db was not getting updated with the actual task exit code
and was recording tasks as failed as a consequence.
With this commit, the snapshot updates are moved so that they occur
after the lock is acquired and before its released so the DB is aware
of the task state when it is updated.