Add task timeout config to "com_scheduler"
Adds a configuration field for a global task timeout.
The timeout serves the purpose of safely declaring a hard-limit on how
long tasks can be considered as running once a lock on them is acquired.
Right now the default is 300s (5min) but this can be re-evaluated
in the future to a more sensible value if needed.