Fix Scheduler::fetchTasks() default behaviour
Shifts the list config for the task queue to the fetchTaskRecord()
method. This makes more sense as the fetchTasks() method is supposed
to be a transparent API method to interact with the TasksModel rather
than have an overriding default like the list.multi_ordering state
variable as a default. Also changes the default ordering behaviour for
sensibility and adds some comments for clarity.