[4.0.x] Update skipto
This updates the skipto js to the newly released 4.1.2

### Version 4.1.2
Added aria-busy="true" attribute to menu element when SkipTo is initialized and being updated with new menu items to support validators looking for required menu items for the menu role.
Added the optional aria-controls attribute to button element to reference the id of the menu element as defined in the W3C ARIA Authoring practices for menu button pattern.

### Version 4.1.1
Removed aria-describedby from button, since screen readers read the accesskey information.