Bugfix scheduleRunner event subscription behaviour
ScheduleRunner attached itself to events even with comScheduler
disabled. This was likely because it utilised a constructor short which
is useful only with the pre-4.x event-method model.
This commit adds checks for the component straight to the
getSubscribedEvents() method and also adds a check for the lazy
scheduler switch in com_scheduler config params.

Refs: @Denitz / joomla/joomla-cms#35143#issuecomment-911453500.