Changes to the class names of info_block.php
Detailed explanation of the changes:

1. On row 17, the class "article-info" is replaced by "article__info" just to follow BEM. 
The class "text-muted" is deleted because we don't want to have this class to every template we use. It is a matter of each template to give different style by using the main class, which in this case was "article-info".
The old class name remained for compatibility reasons, separated by the character | to be indicated as deprecated.

2. On row 23, the class "article-info-term" is replaced by "article__info__title" just to follow BEM.
The old class name remained for compatibility reasons, separated by the character | to be indicated as deprecated.