[4.0] php memory limit
It is possible to set the memory limit in php to -1 which means unlimited memory.

Unfortunately this was not accounted for in the maths to check if a file could be uploaded.

To test you will need to change the value in your php.ini to `memory_limit = -1` and restart the web server.

Then try to upload any file in the media manager and you will see an error message that the file is too large.

Apply this PR and the upload will be successful.

PR for #35853