[4.0] Do not minify tinymce languages
Because of the way that tinymce works it can only use a language file that is named exactly the same as the language. As a result none of the minified files eg de.min.js will ever be used.

They are also listed as a language in the plugin configuration which is obviously also not correct.

this PR stops the minification of these languages files.

To test you  should
1. delete the contents of `media\vendor\tinymce\langs`
2. run `npm run build:js`
3. confirm that there are now no *.min.js in the `media\vendor\tinymce\langs` folder