[4.0] No Articles Message
Both category blog and category list have an option to show or hide a message if there are no articles in the category.

Before this PR the option worked _only_ for a category blog. It was not possible to hide the message for a category list.

In addition the message, when displayed, was not formatted consistently. In some places it was a simple paragraph, in others it was a nicely formatted message. This PR makes the messages display consistently.