[4.0] modals invisible tinymce fullscreen
Tinymce can be made fullscreen by either the hotkey ctrl-shif-f or by clicking on the icon in the toolbar.

This sets the z-index of the editor to 1250. As a result the modals generated by the cms content buttons (editor-xtd pluguins) are invisible as they have a z-index of 1050.

This PR changes the tinymce zindex to 1049