[4.0] header.css unused
To the best of my knowledge and testing ability these classes are no longer used and can be safely removed. It looks like they were missed during the numerous refactoring of the atum template.

When testing either confirm they are not used by code review and/or check that the ... menu is still working as desired on smaller screens