[4.0] tinymce plugin configuration re-paint (#36066)
* [4.0] tinymce plugin configuration re-paint

With J4 we are seeing more people switching to use tinymce. This has just re-emphasised the terrible ui where people do not realise that everything is configured per set. And as the first set that is displayed is set 2 ie not the one for a super admin they make a change to the plugin config and nothing happens.

This PR attempts to go some way to fixing that by
1. switching from bootstrap tabs to the joomla-tab webcomponent
2. changing the order so that set 0 is first and not last

This is really just a subtle change but it appears to make it more obvious that everything is in a set.

* better sort thanks @fedik

* need more coffee

* oops

* tabs