[4.1] Language of Parts [a11y] (#35607)
* [4.1] Language of Parts [a11y]

This PR is a draft to implement and showcase the new language of parts feature in tinyMCE

It has been made to 4.1 as its a new feature

This PR requires #35605 which at the time of this PR has not been merged into 4.1dev

This is a replacement PR of #30939

This wraps the desired text in span tags with a lang attribute for the specified language. Unspecified text is assumed to be written in the page's language. This helps the resulting text comply with WCAG 2.0 3.1.2 Language of Parts: "The human language of each passage or phrase in the content can be programmatically determined..."

If you have customised your tinymce editor toolbar you will need to edit the toolbar again to include this button.

The need for this is based on the EU funded research project for improving the process of creating accessible content by authors https://accessibilitycluster.com/about

You can view a video that demonstrates the benefits of this feature https://www.youtube.com/watch?v=BY9_xhjtLV4 and read the [Technical Sepecification](https://www.dropbox.com/s/mbzh30rdt0c0gqa/Technical%20specification%20-%20Change%20language%20%28We4Authors%20Cluster%29.pdf) that the research project produced.

This PR is only a draft

The remaining task is to decide
- which languages to list
- if they should be translatable
- should they be fr for fr-FR
- or should the list of languages be user selectable in the tinymce plugin configuration

In addition the research project recommended that there should be some form of visible indicator to the content author that a piece of text has been marked as being in a specific language. TinyMCE have not implemented this (yet) but I think we could do it with some css?

cc @chmst

* css

* xml

* lint

* array

* language strings

* Update plugins/editors/tinymce/tinymce.php

Co-authored-by: Quy <quy@fluxbb.org>

Co-authored-by: Quy <quy@fluxbb.org>
Co-authored-by: Benjamin Trenkle <bembelimen@users.noreply.github.com>