[4.0] admin menu RTL
This PR removes code that was required in joomla 3 when the admin menu was horizontal. Now that it is vertical there is no need for this code at all.

To test ensure that the admin menu works and looks the same as before the PR with an RTL language such as arabic or persian