[4.0] media manager infobar
The RTL specific css was wrong as can be seen in the before screenshot -having a seperate block of css for rtl or even worse a separate file makes this sort of thing all too easy to happen.

This PR removes the RTL specific css (should have been scss) and updates the scss to use logical properties instead.

To test RTL you will need to install arabic or persian.

and dont forget to npm ci

There is no visible change in LTR and the RTL changes are as shown below

## Before LTR

## Before RTL

## After LTR

## After RTLFix codestyle