[4.1] Action Logs improved search
This simple PR changes the search in the action logs to include the message field (aka the Action Column) so that now you can find every action that touched a specific article etc.

As this includes the username we can simply change the query from username to message

To test make sure you have more than 1 user that has performed actions on the site. Then make sure that the search finds the results you expect **including** testing the searches with the filters applied