[4.1] Template Override switches
For some reason the switches were not displayed in the same way as other similar switches. This leads to confusion (reported elsewhere) and the need for creating additional css.

This Pr resolves that. Made against 4.1 as there are language string removals and additions and I dont know what the policy is for that in j4.

### Before

### After