Ensure returned array has unique values only
I stumbled over this on a site I was debugging. The system sends the following query:

```
SELECT `folder` AS `type`,`element` AS `name`,`params` AS `params`,`extension_id` AS `id`

  FROM #__extensions

  WHERE enabled = 1 
  AND type = 'plugin' 
  AND state IN (0,1) 
  AND access IN (1,1,5)

  ORDER BY ordering
```

Notice the ```AND access IN (1,1,5)``` the extra 1 comes from the fact that this method hardcodes the 1 on line 1027 and then later also finds in the database. This does not really break anything but is just proper data management. For instance if someone wanted to could how many groups have access to something and used this method and counted the entries would get the answer 3 and not 2. It also leads to "wasted" time with developers who are curious where this comes from and spends time hunting down the source.